+++ internal/cmd/cluster.go
@@ -15,6 +15,7 @@
 package cmd
 
 import (
+	ratesClusters "github.com/sapcc/gophercloud-sapcc/rates/v1/clusters"
 	"github.com/sapcc/gophercloud-sapcc/resources/v1/clusters"
 	"github.com/spf13/cobra"
 
@@ -30,6 +31,7 @@ func newClusterCmd() *cobra.Command {
 	}
 	doNotSortFlags(cmd)
 	cmd.AddCommand(newClusterShowCmd().Command)
+	cmd.AddCommand(newClusterShowRatesCmd().Command)
 	return cmd
 }
 
@@ -88,4 +90,60 @@ func (c *clusterShowCmd) Run(_ *cobra.Command, _ []string) error {
 	}
 
 	return writeReports(outputOpts, core.ClusterReport{ClusterReport: limesRep})
+}
+
+///////////////////////////////////////////////////////////////////////////////
+// Cluster show rates.
+
+type clusterShowRatesCmd struct {
+	*cobra.Command
+
+	rateFilterFlags
+	rateOutputFmtFlags
+}
+
+func newClusterShowRatesCmd() *clusterShowRatesCmd {
+	clusterShowRates := &clusterShowRatesCmd{}
+	cmd := &cobra.Command{
+		Use:     "show-rates",
+		Short:   "Display global rate limits for the cluster",
+		Long:    "Display global rate limits for the cluster level. These rate limits apply to all users in aggregate.",
+		Args:    cobra.NoArgs,
+		PreRunE: authWithLimesRates,
+		RunE:    clusterShowRates.Run,
+	}
+
+	// Flags
+	doNotSortFlags(cmd)
+	clusterShowRates.rateFilterFlags.AddToCmd(cmd)
+	clusterShowRates.rateOutputFmtFlags.AddToCmd(cmd)
+
+	clusterShowRates.Command = cmd
+	return clusterShowRates
+}
+
+func (c *clusterShowRatesCmd) Run(_ *cobra.Command, args []string) error {
+	outputOpts, err := c.rateOutputFmtFlags.validate()
+	if err != nil {
+		return err
+	}
+
+	res := ratesClusters.Get(limesRatesClient, ratesClusters.GetOpts{
+		Areas:    c.areas,
+		Services: c.services,
+	})
+	if res.Err != nil {
+		return util.WrapError(res.Err, "could not get cluster report")
+	}
+
+	if c.format == core.OutputFormatJSON {
+		return writeJSON(res.Body)
+	}
+
+	limesRep, err := res.Extract()
+	if err != nil {
+		return util.WrapError(err, "could not extract cluster report")
+	}
+
+	return writeReports(outputOpts, core.ClusterRatesReport{ClusterReport: limesRep})
 }