+++ CHANGELOG.md
@@ -7,7 +7,17 @@ The sections should follow the order `Added`, `Changed`, `Fixed`, `Removed`, and
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/sapcc/limesctl/compare/v3.1.3...HEAD)
+## [Unreleased](https://github.com/sapcc/limesctl/compare/v3.2.0...HEAD)
+
+## 3.2.0 - 2023-03-03
+
+## Added
+
+- Support for showing cluster global rate limits.
+
+  ```
+  limesctl cluster show-rates
+  ```
 
 ## 3.1.3 - 2022-12-15